Change(timetable): move 'ToDictionary' method to separated class
