+++ Vulcanova/Vulcanova/Features/Dashboard/DashboardViewModel.cs
@@ -207,49 +207,9 @@ namespace Vulcanova.Features.Dashboard
 
             return _timetableService.GetPeriodEntriesByMonth(accountId, monthAndYear, forceSync)
                 .CombineLatest(changes)
-                .Select(items => ToDictionary(items.First, items.Second));
+                .Select(items => TimetableExtensions.ToDictionary(items.First, items.Second));
         }
 
-        private static IReadOnlyDictionary<DateTime, IEnumerable<TimetableListEntry>> ToDictionary(
-            IEnumerable<TimetableEntry> lessons, IEnumerable<TimetableChangeEntry> changes)
-        {
-            var result = new Dictionary<DateTime, IEnumerable<TimetableListEntry>>();
-
-            var groups = lessons.Where(l => l.Visible).GroupBy(e => e.Date.Date);
-
-            // avoid multiple enumerations
-            var timetableChangeEntries = changes as TimetableChangeEntry[] ?? changes.ToArray();
-
-            foreach (var group in groups)
-            {
-                var entries = new List<TimetableListEntry>();
-                foreach (var lesson in group.OrderBy(e => e.Start))
-                {
-                    var change = timetableChangeEntries
-                        .FirstOrDefault(c => c.TimetableEntryId == lesson.Id);
-
-                    var entry = new TimetableListEntry
-                    {
-                        No = lesson.No,
-                        Start = lesson.Start,
-                        Date = lesson.Date,
-                        End = lesson.End,
-                        SubjectName = change?.Subject?.Name ?? lesson.Subject?.Name,
-                        RoomName = change?.RoomName ?? lesson.RoomName,
-                        TeacherName = change?.TeacherName ?? lesson.TeacherName,
-                        Change = change?.Change.Type,
-                        ChangeNote = change?.Note ?? change?.Reason
-                    };
-
-                    entries.Add(entry);
-                }
-
-                result[group.Key] = entries.AsReadOnly();
-            }
-
-            return result;
-        }
-        
         private IObservable<IReadOnlyCollection<Homework.Homework>> GetHomeworkEntries(int accountId, int periodId,
             bool forceSync = false)
         {
